Updated API Key and fixed id from int to long

+++ app/src/main/java/com/example/waterplants/api/model/Suggestion.kt
@@ -6,7 +6,7 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class Suggestion(
     val confirmed: Boolean,
-    val id: Int,
+    val id: Long, // TODO: ZMIENIAM z int na long 12.11
     val plant_details: PlantDetails,
     val plant_name: String,
     val probability: Double,
